mantem somente um atributo com os codigos de rastreio e sempre atribui a ele
